Add files via upload
This python program performs Exploratory Data Analysis  on the existing csv file.  It prints dataset info, summary statistics, and checks for missing values.  
It shows the number of unique values and data types for each column.  It also visualizes the data using histograms to show distributions and boxplots to detect outliers in numerial columns.